641: Renaming com.securebanking.gateway.DcrErrorResponseFactory => com.securebanking.gateway.dcr.ErrorResponseFactory
Issue: https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/641
